RFTR 휴가 신청 로직 중 메시지Q 생성 로직 분리
1. 메시지 생성 로직은 비동기로 처리하며
트랜잭션도 신규로 생성된다. 메시지 생성 실패 시, 최소한의 정보를 얻기 위해 , 예외를 체킹하며 VacationCreatedEvent 정보를 확인한다.

+++ vacation-core/src/main/java/com/jxx/vacation/core/vacation/domain/entity/VacationDuration.java
@@ -39,7 +39,7 @@ public class VacationDuration {
         this.endDateTime = endDateTime;
     }
 
-    public long calculateDate() {
+    public float calculateDate() {
         return ChronoUnit.DAYS.between(startDateTime, endDateTime) + DATE_ADJUSTMENTS_VALUE;
     }
